requirements.txt
nltk
django-admin